Time: 43 ms (21.14%), Space: 34 MB (95.06%) - LeetHub

+++ 2807-insert-greatest-common-divisors-in-linked-list/2807-insert-greatest-common-divisors-in-linked-list.cpp
@@ -16,7 +16,7 @@ public:
         ListNode* cur = head;
         while (cur->next)
         {
-            ListNode* temp = new ListNode(__gcd(cur->val, cur->next->val), cur->next);
+            ListNode* temp = new ListNode((int)__gcd(cur->val, cur->next->val), cur->next);
             cur->next = temp;
             cur = temp->next;
         }
